[Gold IV] Title: 그룹 나누기 (Subset), Time: 0 ms, Memory: 2020 KB, Score: 100 point -BaekjoonHub

+++ 백준/Gold/16132. 그룹 나누기 （Subset）/그룹 나누기 （Subset）.cc
@@ -5,11 +5,11 @@ using namespace std;
 
 int main()
 {
-	int n;
+	long long n;
 	cin >> n;
 
-	int sum = 0;
-	for (int i = 1; i <= n; i++)
+	long long sum = 0;
+	for (long long i = 1; i <= n; i++)
 		sum += i;
 
 	if (sum % 2 == 1)
@@ -20,24 +20,14 @@ int main()
 
 	// 합이 같아야 함
 	// 그렇기에 최종 목표값이 sum / 2
-	int target = sum / 2;
+	long long target = sum / 2;
 
-	// dp 정의
-	// 
-	// - dp[i] ? 
-	//   : i값을 만들기 위한 경우의 수?
-	//
-	// 상태 전이
-	// - dp[i + n[i]] += dp[i]
-	// 
-	//
-
-	vector<int> dp(target + 1, -1);
+	vector<long long> dp(target + 1, -1);
 	dp[0] = 1;
 
-	for (int i = 1; i <= n; i++)
+	for (long long i = 1; i <= n; i++)
 	{
-		for (int j = target; j >= i;j--)
+		for (long long j = target; j >= i;j--)
 		{
 			if (dp[j - i] == -1)
 				continue;
